requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Agents and LLM
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<1.0.0
langchain-community>=0.3.0,<0.4.0
langchain-ollama>=0.1.0

# Face detection
facenet-pytorch>=2.5.3

# OCR
easyocr>=1.7.0

# Object detection
ultralytics>=8.0.0

# Data validation
pydantic>=2.0.0

# Database
pymongo>=4.6.0
cryptography>=41.0.0

# Logging
loguru>=0.7.0

# Progress bars
tqdm>=4.65.0
rich>=13.5.0

# Testing
pytest>=7.4.0
